refactor: split frontend/src/pages/metadata_edit.rs (#314)
Pure file-shape refactor. The 856-line page violated the ~800-line file
cap in .claude/rules/05-rust-style.md. Extracted page-local sub-components
into pages/metadata_edit/{fields,form_grid,header,save_bar,sidebar}.rs;
the page itself is now a slim composition root that still owns every
signal, the dirty-fields memo, and the async save/revert handlers.

No markup or behavior change: every data-testid, label, role, route,
and ChipEditor testid_prefix is preserved. The label_to_id slug tests
travel with their function into fields.rs.

Closes #314
